Fix navigation push and remove cart button
