+++ mobile_app/pubspec.yaml
@@ -20,6 +20,7 @@ dependencies:
   flutter_native_splash: ^2.3.10
   flutter_pdfview: ^1.3.2
   path_provider: ^2.1.1
+  go_router: ^13.2.0
 
 dev_dependencies:
   flutter_test:
